requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alpaca-trade-api==2.3.0
altair==4.2.0
appnope==0.1.3
astroid==2.11.6
asttokens==2.0.8
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
base58==2.1.1
bitarray==2.6.0
black==22.3.0
bleach==5.0.1
blinker==1.5
bokeh==2.4.3
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorcet==3.0.0
commonmark==0.9.1
cycler==0.11.0
cytoolz==0.12.0
decorator==5.1.1
deprecation==2.1.0
deta==1.1.0
dill==0.3.5.1
docutils==0.19
entrypoints==0.4
enum34==1.1.10
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.1
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
executing==1.1.0
fire==0.3.1
flake8==4.0.1
Flask==2.2.2
fonttools==4.34.4
frozenlist==1.3.0
gitdb==4.0.9
GitPython==3.1.29
greenlet==1.1.2
hexbytes==0.3.0
holoviews==1.15.0
hvplot==0.8.0
idna==3.3
importlib-metadata==4.12.0
iniconfig==1.1.1
ipfshttpclient==0.8.0a2
ipython==8.5.0
isort==5.10.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.17.0
keyring==23.8.2
kiwisolver==1.4.4
lazy-object-proxy==1.7.1
lru-dict==1.1.8
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.6
mccabe==0.6.1
msgpack==1.0.3
multiaddr==0.0.9
multidict==6.0.2
mypy-extensions==0.4.3
netaddr==0.8.0
networkx==2.8.8
numpy==1.23.1
packaging==21.3
pandas==1.4.3
panel==0.13.1
param==1.12.2
parsimonious==0.8.1
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pkginfo==1.8.3
platformdirs==2.5.2
pluggy==1.0.0
prompt-toolkit==3.0.29
protobuf==3.19.5
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyarrow==7.0.0
pycodestyle==2.8.0
pycryptodome==3.15.0
pyct==0.4.8
pydeck==0.8.0
pyflakes==2.4.0
Pygments==2.12.0
pylint==2.14.3
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.2
pytest==7.1.3
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyviz-comms==2.2.0
PyYAML==6.0
questionary==1.5.2
readme-renderer==36.0
requests==2.28.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
rich==12.5.1
rlp==2.0.1
semver==2.13.0
six==1.16.0
smmap==5.0.0
SQLAlchemy==1.4.39
stack-data==0.5.1
streamlit==1.15.1
streamlit-lottie==0.0.3
streamlit-option-menu==0.3.2
termcolor==1.1.0
tokenize-rt==5.0.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.0
toolz==0.12.0
tornado==6.2
tqdm==4.64.0
traitlets==5.4.0
twine==4.0.1
typing_extensions==4.2.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.10
validators==0.20.0
varint==1.0.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
websockets==9.1
Werkzeug==2.2.2
wrapt==1.14.1
xarray==2022.6.0
yarl==1.7.2
zipp==3.8.1
